added flashed messages

+++ app/routes/events.js
@@ -3,7 +3,7 @@ const router = express.Router()
 const client = require('../db.js')
 const {getEventByPublicId, deleteEventByPublicId, getAllEvents} = require('../models/event')
 const auth = require('../auth')
-const {flashMsg} = require('../flashMessages.js')
+const flashMsg = require('../flashMessages.js')
 
 
 router.get("/", auth.checkAuthenticated,  async (req, res) => {
@@ -11,7 +11,7 @@ router.get("/", auth.checkAuthenticated,  async (req, res) => {
         var events = await getAllEvents(req.user.id)        
     } catch (e) {
         console.error(e)
-        flashMsg.generalError(req)
+        req.flash(flashMsg.generalError.htmlClass, flashMsg.generalError.msg)
         return res.redirect('events')
     } 
     return res.render('event/events', {events: events, isAuthenticated: true}) 
@@ -19,6 +19,10 @@ router.get("/", auth.checkAuthenticated,  async (req, res) => {
 
 router.get("/:public_id", auth.checkAuthenticated,  async (req, res) => {
     const event = await getEventByPublicId(req.params.public_id)
+    if (event.user_id != req.user.id) {
+        req.flash(flashMsg.insufficientPrivileges.htmlClass, flashMsg.insufficientPrivileges.msg)
+        return res.redirect('/events')
+    }
     
     if (event.length === 0) {
         req.flash('error', 'This event does not exist')
@@ -29,9 +33,15 @@ router.get("/:public_id", auth.checkAuthenticated,  async (req, res) => {
 
 router.delete("/:public_id", auth.checkAuthenticated,  async (req, res) => {
     try {
+        const event = await getEventByPublicId(req.params.public_id)
+        if (event.user_id != req.user.id) {
+            req.flash(flashMsg.insufficientPrivileges.htmlClass, flashMsg.insufficientPrivileges.msg)
+            return res.redirect('/events')
+        }
         await deleteEventByPublicId(req.params.public_id)
+        console.log(`Event with public id ${req.params.public_id} deleted successfully`)
     } catch (e) {
-        flashMsg.generalError(req)
+        req.flash(flashMsg.generalError.htmlClass, flashMsg.generalError.msg)
         return res.redirect('/events')
     }
     req.flash('success', 'Event deleted successfully')
